[DITP-pilotage/pilote-2] refactor: factorise les primitifs UI restants dans kpilote-ui
Déplace dans le package tous les primitifs génériques restants :
- webapp : BackLink, Card, CardGrid, Collapsible, DataTable, DescriptionList,
  DropdownMenu, EntityCard, Page, Pagination, Pill, ProgressBar, Section,
  StatCard, StatGrid, SearchField.
- admin : CopyButton, FadeIn.

La map exports du package est régénérée depuis src/. Restent app-spécifiques :
JsonEditor (CodeMirror/console), BarCard (nav dashboard admin), ResponsablesList
(métier responsable), et Tabs/TabNav (à réconcilier ensuite).

Refs:
Project: pilote-2
Client: DITP-pilotage
Category: refactor
Stack: react, typescript
Tags: design-system, monorepo
Files-changed: 44
