+++ apps/kpilote-webapp/src/components/RouteError.tsx
@@ -2,10 +2,10 @@ import type { ErrorComponentProps } from '@tanstack/react-router'
 import { Link } from '@tanstack/react-router'
 import { HTTPError } from 'ky'
 
-import { BackLink } from '@/components/ui/BackLink'
+import { BackLink } from '@pilote/kpilote-ui/BackLink'
 import { Button } from '@pilote/kpilote-ui/Button'
 import { EmptyState } from '@pilote/kpilote-ui/EmptyState'
-import { Section } from '@/components/ui/Section'
+import { Section } from '@pilote/kpilote-ui/Section'
 
 const isNotFoundError = (error: Error): boolean => {
   return error instanceof HTTPError && error.response.status === 404
