+++ apps/kpilote-webapp/src/components/paniers/PanierAvancement.tsx
@@ -1,6 +1,6 @@
 import { useSuspenseQuery } from '@tanstack/react-query'
 
-import { ProgressBar } from '@/components/ui/ProgressBar'
+import { ProgressBar } from '@pilote/kpilote-ui/ProgressBar'
 import { Text } from '@pilote/kpilote-ui/Typography'
 import { formatNumberFr } from '@/lib/format'
 import { panierTauxProgressionIndividuQueryOptions } from '@/queries/panierTauxProgression'
