+++ apps/kpilote-webapp/src/routes/_authenticated/indicateurs/$id.tsx
@@ -14,11 +14,11 @@ import { IndicateurResultatsTab } from '@/components/indicateurs/IndicateurResul
 import { IndividuSelect } from '@/components/indicateurs/IndividuSelect'
 import { useImportModal } from '@/components/import-valeurs/useImportModal'
 import { usePageFileDrop } from '@/components/import-valeurs/usePageFileDrop'
-import { BackLink } from '@/components/ui/BackLink'
+import { BackLink } from '@pilote/kpilote-ui/BackLink'
 import { Button } from '@pilote/kpilote-ui/Button'
 import { EmptyState } from '@pilote/kpilote-ui/EmptyState'
 import { Field } from '@pilote/kpilote-ui/Field'
-import { Page } from '@/components/ui/Page'
+import { Page } from '@pilote/kpilote-ui/Page'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/Tabs'
 import { ensureIndividuReferentielPair } from '@/lib/individus/pair'
 import { useRecordVisit } from '@/lib/recentlyVisited'
