+++ apps/kpilote-webapp/src/routes/_authenticated/indicateurs/index.tsx
@@ -10,14 +10,14 @@ import { RouteError } from '@/components/RouteError'
 import { RouteLoading } from '@/components/RouteLoading'
 import { IndicateurCard } from '@/components/indicateurs/IndicateurCard'
 import { IndividuSelect } from '@/components/indicateurs/IndividuSelect'
-import { CardGrid } from '@/components/ui/CardGrid'
+import { CardGrid } from '@pilote/kpilote-ui/CardGrid'
 import { EmptyState } from '@pilote/kpilote-ui/EmptyState'
 import { Field } from '@pilote/kpilote-ui/Field'
-import { Page } from '@/components/ui/Page'
-import { SearchField } from '@/components/ui/SearchField'
+import { Page } from '@pilote/kpilote-ui/Page'
+import { SearchField } from '@pilote/kpilote-ui/SearchField'
 import { Text } from '@pilote/kpilote-ui/Typography'
 import { ensureIndividuReferentielPair } from '@/lib/individus/pair'
-import { DEFAULT_PAGE_SIZE_OPTIONS, Pagination } from '@/components/ui/Pagination'
+import { DEFAULT_PAGE_SIZE_OPTIONS, Pagination } from '@pilote/kpilote-ui/Pagination'
 import { indicateursQueryOptions, loadIndicateurs } from '@/queries/indicateurs'
 import { allReferentielsQueryOptions, loadAllReferentielIds } from '@/queries/referentiels'
