+++ apps/kpilote-webapp/src/routes/_authenticated/paniers/$id.tsx
@@ -14,11 +14,11 @@ import { IndividuSelect } from '@/components/indicateurs/IndividuSelect'
 import { PanierCommentaireConfigProvider } from '@/components/paniers/PanierCommentaireConfigProvider'
 import { PanierGouvernanceTab } from '@/components/paniers/PanierGouvernanceTab'
 import { PanierTauxProgression } from '@/components/paniers/PanierTauxProgression'
-import { BackLink } from '@/components/ui/BackLink'
-import { CardGrid } from '@/components/ui/CardGrid'
+import { BackLink } from '@pilote/kpilote-ui/BackLink'
+import { CardGrid } from '@pilote/kpilote-ui/CardGrid'
 import { EmptyState } from '@pilote/kpilote-ui/EmptyState'
 import { Field } from '@pilote/kpilote-ui/Field'
-import { Page } from '@/components/ui/Page'
+import { Page } from '@pilote/kpilote-ui/Page'
 import { Tabs, TabsContent, TabsList, TabsTrigger } from '@/components/ui/Tabs'
 import { Text } from '@pilote/kpilote-ui/Typography'
 import { ensureIndividuReferentielPair } from '@/lib/individus/pair'
