+++ apps/kpilote-webapp/src/routes/_authenticated/paniers/index.tsx
@@ -10,11 +10,11 @@ import { RouteError } from '@/components/RouteError'
 import { RouteLoading } from '@/components/RouteLoading'
 import { IndividuSelect } from '@/components/indicateurs/IndividuSelect'
 import { PanierCard } from '@/components/paniers/PanierCard'
-import { CardGrid } from '@/components/ui/CardGrid'
+import { CardGrid } from '@pilote/kpilote-ui/CardGrid'
 import { EmptyState } from '@pilote/kpilote-ui/EmptyState'
 import { Field } from '@pilote/kpilote-ui/Field'
-import { Page } from '@/components/ui/Page'
-import { DEFAULT_PAGE_SIZE_OPTIONS, Pagination } from '@/components/ui/Pagination'
+import { Page } from '@pilote/kpilote-ui/Page'
+import { DEFAULT_PAGE_SIZE_OPTIONS, Pagination } from '@pilote/kpilote-ui/Pagination'
 import { Text } from '@pilote/kpilote-ui/Typography'
 import { ensureIndividuReferentielPair } from '@/lib/individus/pair'
 import { loadPaniers, paniersQueryOptions } from '@/queries/paniers'
