+++ packages/kpilote-ui/package.json
@@ -5,57 +5,57 @@
   "type": "module",
   "exports": {
     "./theme.css": "./theme.css",
-    "./clsxm": {
-      "types": "./src/clsxm.ts",
-      "default": "./src/clsxm.ts"
+    "./BackLink": {
+      "types": "./src/BackLink.tsx",
+      "default": "./src/BackLink.tsx"
     },
     "./Button": {
       "types": "./src/Button.tsx",
       "default": "./src/Button.tsx"
     },
-    "./Typography": {
-      "types": "./src/Typography.tsx",
-      "default": "./src/Typography.tsx"
+    "./Card": {
+      "types": "./src/Card.tsx",
+      "default": "./src/Card.tsx"
     },
-    "./EmptyState": {
-      "types": "./src/EmptyState.tsx",
-      "default": "./src/EmptyState.tsx"
+    "./CardGrid": {
+      "types": "./src/CardGrid.tsx",
+      "default": "./src/CardGrid.tsx"
     },
-    "./Marianne": {
-      "types": "./src/Marianne.tsx",
-      "default": "./src/Marianne.tsx"
+    "./Collapsible": {
+      "types": "./src/Collapsible.tsx",
+      "default": "./src/Collapsible.tsx"
     },
-    "./Toast": {
-      "types": "./src/Toast.tsx",
-      "default": "./src/Toast.tsx"
+    "./CopyButton": {
+      "types": "./src/CopyButton.tsx",
+      "default": "./src/CopyButton.tsx"
+    },
+    "./DataTable": {
+      "types": "./src/DataTable.tsx",
+      "default": "./src/DataTable.tsx"
+    },
+    "./DescriptionList": {
+      "types": "./src/DescriptionList.tsx",
+      "default": "./src/DescriptionList.tsx"
     },
     "./Dialog": {
       "types": "./src/Dialog.tsx",
       "default": "./src/Dialog.tsx"
     },
-    "./Modale": {
-      "types": "./src/Modale.tsx",
-      "default": "./src/Modale.tsx"
+    "./DropdownMenu": {
+      "types": "./src/DropdownMenu.tsx",
+      "default": "./src/DropdownMenu.tsx"
     },
-    "./ModaleForm": {
-      "types": "./src/ModaleForm.tsx",
-      "default": "./src/ModaleForm.tsx"
+    "./EmptyState": {
+      "types": "./src/EmptyState.tsx",
+      "default": "./src/EmptyState.tsx"
     },
-    "./Picker": {
-      "types": "./src/Picker.tsx",
-      "default": "./src/Picker.tsx"
+    "./EntityCard": {
+      "types": "./src/EntityCard.tsx",
+      "default": "./src/EntityCard.tsx"
     },
-    "./PickerOptionNomId": {
-      "types": "./src/PickerOptionNomId.tsx",
-      "default": "./src/PickerOptionNomId.tsx"
-    },
-    "./SegmentedControl": {
-      "types": "./src/SegmentedControl.tsx",
-      "default": "./src/SegmentedControl.tsx"
-    },
-    "./Table": {
-      "types": "./src/Table.tsx",
-      "default": "./src/Table.tsx"
+    "./FadeIn": {
+      "types": "./src/FadeIn.tsx",
+      "default": "./src/FadeIn.tsx"
     },
     "./Field": {
       "types": "./src/Field.tsx",
@@ -73,13 +73,85 @@
       "types": "./src/FieldTextarea.tsx",
       "default": "./src/FieldTextarea.tsx"
     },
-    "./Select": {
-      "types": "./src/Select.tsx",
-      "default": "./src/Select.tsx"
+    "./Marianne": {
+      "types": "./src/Marianne.tsx",
+      "default": "./src/Marianne.tsx"
+    },
+    "./Modale": {
+      "types": "./src/Modale.tsx",
+      "default": "./src/Modale.tsx"
+    },
+    "./ModaleForm": {
+      "types": "./src/ModaleForm.tsx",
+      "default": "./src/ModaleForm.tsx"
+    },
+    "./Page": {
+      "types": "./src/Page.tsx",
+      "default": "./src/Page.tsx"
+    },
+    "./Pagination": {
+      "types": "./src/Pagination.tsx",
+      "default": "./src/Pagination.tsx"
+    },
+    "./Picker": {
+      "types": "./src/Picker.tsx",
+      "default": "./src/Picker.tsx"
+    },
+    "./PickerOptionNomId": {
+      "types": "./src/PickerOptionNomId.tsx",
+      "default": "./src/PickerOptionNomId.tsx"
+    },
+    "./Pill": {
+      "types": "./src/Pill.tsx",
+      "default": "./src/Pill.tsx"
+    },
+    "./ProgressBar": {
+      "types": "./src/ProgressBar.tsx",
+      "default": "./src/ProgressBar.tsx"
+    },
+    "./SearchField": {
+      "types": "./src/SearchField.tsx",
+      "default": "./src/SearchField.tsx"
+    },
+    "./Section": {
+      "types": "./src/Section.tsx",
+      "default": "./src/Section.tsx"
+    },
+    "./SegmentedControl": {
+      "types": "./src/SegmentedControl.tsx",
+      "default": "./src/SegmentedControl.tsx"
     },
     "./SegmentedField": {
       "types": "./src/SegmentedField.tsx",
       "default": "./src/SegmentedField.tsx"
+    },
+    "./Select": {
+      "types": "./src/Select.tsx",
+      "default": "./src/Select.tsx"
+    },
+    "./StatCard": {
+      "types": "./src/StatCard.tsx",
+      "default": "./src/StatCard.tsx"
+    },
+    "./StatGrid": {
+      "types": "./src/StatGrid.tsx",
+      "default": "./src/StatGrid.tsx"
+    },
+    "./Table": {
+      "types": "./src/Table.tsx",
+      "default": "./src/Table.tsx"
+    },
+    "./Toast": {
+      "types": "./src/Toast.tsx",
+      "default": "./src/Toast.tsx"
+    },
+    "./Typography": {
+      "types": "./src/Typography.tsx",
+      "default": "./src/Typography.tsx"
+    },
+    "./clsxm": {
+      "types": "./src/clsxm.ts",
+      "default": "./src/clsxm.ts"
     }
   },
   "files": [
