+++ packages/kpilote-ui/src/EntityCard.tsx
@@ -2,8 +2,8 @@ import { Slot as SlotPrimitive } from 'radix-ui'
 import { ArrowUpRight } from 'lucide-react'
 import type { ComponentProps, ReactNode } from 'react'
 
-import { Heading, Text } from '@pilote/kpilote-ui/Typography'
-import { clsxm } from '@/lib/clsxm'
+import { Heading, Text } from './Typography'
+import { clsxm } from './clsxm'
 
 type EntityCardProps = Omit<ComponentProps<'div'>, 'title'> & {
   asChild?: boolean
