+++ packages/kpilote-ui/src/StatGrid.tsx
@@ -1,6 +1,6 @@
 import type { ReactNode } from 'react'
 
-import { clsxm } from '@/lib/clsxm'
+import { clsxm } from './clsxm'
 
 type StatGridProps = {
   columns?: 2 | 3 | 4
